include: sbi: Support CHERI capability in struct sbi_context
Use uintptr_t for stvec, sscratch and sepc for Zcheripurecap ABI compatible.

+++ include/sbi/sbi_domain_context.h
@@ -21,11 +21,11 @@ struct sbi_context {
 	/** Supervisor interrupt enable register */
 	unsigned long sie;
 	/** Supervisor trap vector base address register */
-	unsigned long stvec;
+	uintptr_t stvec;
 	/** Supervisor scratch register for temporary storage */
-	unsigned long sscratch;
+	uintptr_t sscratch;
 	/** Supervisor exception program counter register */
-	unsigned long sepc;
+	uintptr_t sepc;
 	/** Supervisor cause register */
 	unsigned long scause;
 	/** Supervisor trap value register */
